requirements.txt
allure-pytest==2.15.3
allure-python-commons==2.15.3
python-dotenv==1.2.1
requests==2.32.5
selenium==4.40.0
typing_extensions==4.15.0
pytest==9.0.2
